[server] | refactor(): 상대 경로로 변경
