+++ packages/server/src/user/dto/user-video-list-response.dto.ts
@@ -1,6 +1,6 @@
 import { UserVideoResponseDto } from './user-video-response.dto';
-import { Video } from '../../../../typeorm/src/entity/video.entity';
-import { UserVideoInfo } from 'user/model/user-video-info';
+import { Video } from '../../../entity/video.entity';
+import { UserVideoInfo } from '../../user/model/user-video-info';
 
 export class UserVideoListResponseDto {
   public constructor(videoList: Video[], count: number) {
